Post 목록 조회 및 검색 구현

+++ seboard/src/main/java/com/seproject/seboard/persistence/CommentSearchJpaRepository.java
@@ -15,5 +15,5 @@ public interface CommentSearchJpaRepository extends CommentSearchRepository {
     @Query("select r from Reply r where r.superComment.commentId = :commentId")
     List<Reply> findReplyListByCommentId(Long commentId);
     @Query("select count(r) from Reply r where r.post.postId = :postId")
-    Long countReplyByPostId(Long postId);
+    int countReplyByPostId(Long postId);
 }
